requirements.txt
PyYAML==5.3.1
pathspec>=0.5.3